Moved LruCache to prevent conflicts.
