+++ server/instrumentation-backend/src/main/java/sh/calaba/instrumentationbackend/query/ast/UIQueryASTWith.java
@@ -15,7 +15,6 @@ import org.antlr.runtime.tree.CommonTree;
 import sh.calaba.instrumentationbackend.actions.webview.CalabashChromeClient;
 import sh.calaba.instrumentationbackend.actions.webview.QueryHelper;
 
-import android.support.v4.util.LruCache;
 import android.view.View;
 
 import sh.calaba.instrumentationbackend.query.WebContainer;
